Removed depricated ResourceViewModel Moved filereader method to FileHandler
